Week-07-Data-Warehouse-Big-Data/olist_dwh_project/requirements.txt
pandas
SQLAlchemy
psycopg2-binary
python-dotenv
matplotlib
seaborn